feat: 3-step insert wizard, single feature list, price-first pricing card (OSS-9, OSS-23, OSS-19)
